Trait mssql_client::Command

pub trait Command {
    fn execute<'a, S, P>(
        self, 
        sql: S, 
        params: P
    ) -> LocalBoxFuture<'a, Result<Self>>
    where
        P: Debug + Params<'a> + 'a,
        S: Debug + Into<Cow<'static, str>> + 'a,
        Self: Sized;

    fn query_fold<'a, T, S, P, F>(
        self, 
        sql: S, 
        params: P, 
        init: T, 
        func: F
    ) -> LocalBoxFuture<'a, Result<(Self, T)>>
    where
        F: FnMut(T, &Row) -> Result<T> + 'a,
        P: Debug + Params<'a> + 'a,
        S: Debug + Into<Cow<'static, str>> + 'a,
        Self: Sized,
        T: 'a;

    fn query<'a, T, S, P>(
        self, 
        sql: S, 
        params: P
    ) -> LocalBoxFuture<'a, Result<(Self, Vec<T>)>>
    where
        P: Debug + Params<'a> + 'a,
        S: Debug + Into<Cow<'static, str>> + 'a,
        Self: Sized,
        T: FromRow + 'a,
    { ... }

    fn query_map<'a, T, S, P, F>(
        self, 
        sql: S, 
        params: P, 
        func: F
    ) -> LocalBoxFuture<'a, Result<(Self, Vec<T>)>>
    where
        F: FnMut(&Row) -> Result<T> + 'a,
        P: Debug + Params<'a> + 'a,
        S: Debug + Into<Cow<'static, str>> + 'a,
        Self: Sized,
        T: 'a,
    { ... }
}

Required methods

fn execute<'a, S, P>(
self,
sql: S,
params: P
) -> LocalBoxFuture<'a, Result<Self>> where
P: Debug + Params<'a> + 'a,
S: Debug + Into<Cow<'static, str>> + 'a,
Self: Sized,

Execute an sql command that does not returns rows.

Example

use futures::Future;
use mssql_client::{Connection, Command, Result};

#[tokio::main]
async fn main() -> Result<()> {
    let conn = Connection::from_env("MSSQL_DB").await?;
    Command::execute(conn, "DECLARE @i INT = @p1", 10).await?;
    Ok(())
}

fn query_fold<'a, T, S, P, F>(
self,
sql: S,
params: P,
init: T,
func: F
) -> LocalBoxFuture<'a, Result<(Self, T)>> where
F: FnMut(T, &Row) -> Result<T> + 'a,
P: Debug + Params<'a> + 'a,
S: Debug + Into<Cow<'static, str>> + 'a,
Self: Sized,
T: 'a,

Loading content...
Provided methods

fn query<'a, T, S, P>(
self,
sql: S,
params: P
) -> LocalBoxFuture<'a, Result<(Self, Vec<T>)>> where
P: Debug + Params<'a> + 'a,
S: Debug + Into<Cow<'static, str>> + 'a,
Self: Sized,
T: FromRow + 'a,

Query the database and reads all rows.

Example

use mssql_client::{Connection, Command, Result};

#[tokio::main]
async fn main() -> Result<()> {
    let conn = Connection::from_env("MSSQL_DB").await?;
    let (_, rows) = Command::query(conn, "SELECT @p1 + 2", 10).await?;

    assert_eq!(12, rows[0]);
    Ok(())
}

fn query_map<'a, T, S, P, F>(
self,
sql: S,
params: P,
func: F
) -> LocalBoxFuture<'a, Result<(Self, Vec<T>)>> where
F: FnMut(&Row) -> Result<T> + 'a,
P: Debug + Params<'a> + 'a,
S: Debug + Into<Cow<'static, str>> + 'a,
Self: Sized,
T: 'a,

Query the database and reads all rows using a function to transform them.

Example

use mssql_client::{Connection, Command, Result};

#[tokio::main]
async fn main() -> Result<()> {
    let conn = Connection::from_env("MSSQL_DB").await?;

    let (_, rows) = Command::query_map(
        conn,
        "SELECT @p1 + 2",
        10,
        |row| row.get(0),
    ).await?;

    assert_eq!(12, rows[0]);
    Ok(())
}

Loading content...