Trait mssql_client::FromColumn

pub trait FromColumn<'a>: Sized {
    type Value: SqlValue<'a>;
    fn from_column(v: Self::Value) -> Result<Self>;
}

This trait convert a sql column value into a rust type. Implement this trait to be able to support more types as needed.

use mssql_client::{FromColumn, Result};

enum Code {
    A,
    B,
}

impl<'a> FromColumn<'a> for Code {
    type Value = &'a str;

    fn from_column(v: Self::Value) -> Result<Self> {
        Ok(match v {
            "A" => Code::A,
            "B" => Code::B,
            _ => return Err("invalid code".into()),
        })
    }
}

Associated Types

type Value: SqlValue<'a>

Loading content...
Required methods

fn from_column(v: Self::Value) -> Result<Self>

Loading content...